# inverted the relationship between expanded macros and tokens