--- a/simplecpp.h
+++ b/simplecpp.h
@@ -172,16 +172,6 @@ namespace simplecpp {
             return tok;
         }
 
-        void setExpandedFrom(const Token *tok, const Macro* m) {
-            mExpandedFrom = tok->mExpandedFrom;
-            mExpandedFrom.insert(m);
-            if (tok->whitespaceahead)
-                whitespaceahead = true;
-        }
-        bool isExpandedFrom(const Macro* m) const {
-            return mExpandedFrom.find(m) != mExpandedFrom.end();
-        }
-
         void printAll() const;
         void printOut() const;
     private: