Instead of duplicating the macro references into each token we should just keep references to the tokens in the macros.