@firewave 2023-08-21
Currently just a hack but tests still pass. And this fixes https://trac.cppcheck.net/ticket/11885.