@firewave 2023-08-24
These changes are better. They pass all the tests in simplecpp and Cppcheck. They do not fix that upstream issue as it hits something different now.

Before
```
Benchmark 1: ./simplecpp -q file.c
  Time (mean ± σ):     433.6 ms ±   4.2 ms    [User: 333.4 ms, System: 99.9 ms]
  Range (min … max):   428.3 ms … 439.6 ms    5 runs
```

After
```
Benchmark 1: ./simplecpp -q file.c
  Time (mean ± σ):     272.8 ms ±   4.8 ms    [User: 223.9 ms, System: 48.5 ms]
  Range (min … max):   266.9 ms … 278.7 ms    5 runs
```

Clang 15 `3,048,566,387` -> `1,328,447,289`