@firewave 2023-08-24
Unfortunately these changes case a big regression to the example in #96.

Before
```
Benchmark 1: ./simplecpp -q file2.c
  Time (mean ± σ):      1.088 s ±  0.033 s    [User: 0.784 s, System: 0.304 s]
  Range (min … max):    1.061 s …  1.125 s    5 runs
```

After 
```
Benchmark 1: ./simplecpp -q file2.c
  Time (mean ± σ):      4.029 s ±  0.061 s    [User: 3.292 s, System: 0.734 s]
  Range (min … max):    3.986 s …  4.134 s    5 runs
```